Fix test clang-tidy warnings: int precision, unused var, emplace_back
- connection_widget_tests.cpp: cast port arithmetic to uint16_t to
  avoid implicit-int-conversion precision loss
- ipc_client_manager_tests.cpp: mark unused result with
  [[maybe_unused]], use emplace_back instead of push_back

+++ tests/unit/connection_widget_tests.cpp
@@ -361,7 +361,7 @@ TEST_F(ConnectionWidgetTest, UpdateMetricsRapidly) {
 TEST_F(ConnectionWidgetTest, SessionInfoUpdatesRapidly) {
   for (int i = 0; i < 50; ++i) {
     widget_->setSessionId(QString("session-%1").arg(i));
-    widget_->setServerAddress(QString("server-%1.example.com").arg(i), 4433 + i);
+    widget_->setServerAddress(QString("server-%1.example.com").arg(i), static_cast<uint16_t>(4433 + i));
   }
 }
